+++ vulnerablecode/urls.py
@@ -18,9 +18,9 @@ from rest_framework.routers import DefaultRouter
 
 from vulnerabilities.api import AliasViewSet
 from vulnerabilities.api import CPEViewSet
-from vulnerabilities.api import PackageV2ViewSet
+from vulnerabilities.v2_api import PackageV2ViewSet
 from vulnerabilities.api import PackageViewSet
-from vulnerabilities.api import VulnerabilityV2ViewSet
+from vulnerabilities.v2_api import VulnerabilityV2ViewSet
 from vulnerabilities.api import VulnerabilityViewSet
 from vulnerabilities.views import ApiUserCreateView
 from vulnerabilities.views import HomePage
